requirements.txt
absl-py==2.1.0
aiofiles==23.2.1
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
anyio==4.11.0
astroid==3.2.4
async-timeout==4.0.3
attrs==24.2.0
blinker==1.8.2
BlurWindow==1.2.1
Brotli==1.1.0
cachetools==5.5.0
Cerberus==1.3.4
certifi==2024.7.4
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.0
cryptography==46.0.3
cssutils==2.5.1
cycler==0.12.1
DataProperty==1.0.1
dill==0.3.8
distro==1.9.0
docker-pycreds==0.4.0
elastic-transport==8.15.0
elasticsearch==8.15.0
exceptiongroup==1.2.2
fastapi==0.112.2
ffmpy==0.4.0
filelock==3.15.4
Flask==3.0.3
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
gekko==1.2.1
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.15
google-api-core==2.19.2
google-api-python-client==2.143.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-genai==1.47.0
google-generativeai==0.8.5
googleapis-common-protos==1.65.0
gradio_client==0.15.1
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httptools==0.7.1
httpx==0.28.1
huggingface-hub==0.36.0
humanfriendly==10.0
humanize==4.3.0
idna==3.7
importlib_metadata==8.4.0
importlib_resources==6.4.4
iniconfig==2.0.0
isort==5.13.2
itsdangerous==2.2.0
Janome==0.5.0
Jinja2==3.1.6
jiter==0.5.0
joblib==1.4.2
jsonlines==4.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
llama_stack==0.0.1a5
llvmlite==0.42.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.4
mbstrdecoder==1.1.3
mccabe==0.7.0
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
mutagen==1.47.0
narwhals==1.6.0
networkx==3.2.1
nltk==3.9.1
numexpr==2.10.1
numpy==2.0.2
orjson==3.10.7
outcome==1.3.0.post0
packaging==24.1
pandas==2.3.3
pathvalidate==3.2.1
pdfminer.six==20251107
pdfplumber==0.11.8
pefile==2024.8.26
pillow==11.3.0
platformdirs==4.2.2
pluggy==1.5.0
portalocker==2.10.1
proto-plus==1.24.0
protobuf==5.29.5
psutil==5.9.1
py2exe==0.13.0.2
pyarrow==17.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pybind11==2.13.5
pycountry==24.6.1
pycparser==2.23
pycryptodomex==3.20.0
pydantic==2.12.3
pydantic_core==2.41.4
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.1.4
pypdfium2==5.3.0
pyperclip==1.11.0
PyQt6==6.7.0
PyQt6-Qt6==6.7.3
PyQt6_sip==13.10.2
pyreadline3==3.4.1
PySocks==1.7.1
pytablewriter==1.2.0
pytest-asyncio==0.21.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-multipart==0.0.9
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
pywin32==304
PyYAML==6.0
referencing==0.35.1
regex==2024.7.24
requests==2.32.5
rouge==1.0.1
rouge_score==0.1.2
rpds-py==0.20.0
rsa==4.9
ruff==0.12.8
sacrebleu==1.5.0
safetensors==0.7.0
scikit-learn==1.5.1
scipy==1.13.1
selenium==4.36.0
semantic-version==2.10.0
sentence-transformers==5.1.2
sentry-sdk==2.13.0
setproctitle==1.3.3
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
sqlitedict==2.1.0
starlette==0.38.2
sympy==1.14.0
tabledata==1.3.3
tcolorpy==0.1.6
tenacity==9.1.2
tensorboard-data-server==0.7.2
termcolor==2.4.0
threadpoolctl==3.5.0
tokenizers==0.22.2
tomli==2.0.1
tomlkit==0.12.0
torch==2.8.0
tqdm==4.67.1
tqdm-multiprocess==0.0.11
transformers==4.57.5
trio==0.31.0
trio-websocket==0.12.2
typepy==1.3.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2024.1
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.19
uvicorn==0.39.0
watchdog==2.1.9
watchfiles==1.1.1
webdriver-manager==4.0.2
websocket-client==1.9.0
websockets==13.0.1
Werkzeug==3.0.4
winsdk==1.0.0b6
wsproto==1.2.0
xxhash==3.5.0
yarl==1.9.4
zipp==3.20.1
zstandard==0.23.0
